chore(clayui.com): move code editor into its own folder
