+++ clayui.com/src/components/clay/Badge.js
@@ -6,7 +6,7 @@
 import ClayBadge from '@clayui/badge';
 import React from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
