+++ clayui.com/src/components/clay/Button.js
@@ -6,7 +6,7 @@
 import ClayButton from '@clayui/button';
 import React from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
