+++ clayui.com/src/components/clay/Card.js
@@ -15,7 +15,7 @@ import ClayLabel from '@clayui/label';
 import ClaySticker from '@clayui/sticker';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
