+++ clayui.com/src/components/clay/ColorPicker.js
@@ -6,7 +6,7 @@
 import ClayColorPicker from '@clayui/color-picker';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
