+++ clayui.com/src/components/clay/DualListBox.js
@@ -6,7 +6,7 @@
 import {ClayDualListBox} from '@clayui/form';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
