+++ clayui.com/src/components/clay/Icon.js
@@ -6,7 +6,7 @@
 import ClayIcon, {ClayIconSpriteContext} from '@clayui/icon';
 import React from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
