+++ clayui.com/src/components/clay/Label.js
@@ -6,7 +6,7 @@
 import ClayLabel from '@clayui/label';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
