+++ clayui.com/src/components/clay/List.js
@@ -6,7 +6,7 @@
 import ClayList from '@clayui/list';
 import React from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
