+++ clayui.com/src/components/clay/ManagementToolbar.js
@@ -13,7 +13,7 @@ import ClayManagementToolbar, {
 } from '@clayui/management-toolbar';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
