+++ clayui.com/src/components/clay/PaginationBar.js
@@ -11,7 +11,7 @@ import ClayPaginationBar, {
 } from '@clayui/pagination-bar';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
