+++ clayui.com/src/components/clay/Panel.js
@@ -6,7 +6,7 @@
 import ClayPanel from '@clayui/panel';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
