+++ clayui.com/src/components/clay/ProgressBar.js
@@ -6,7 +6,7 @@
 import ClayProgressBar from '@clayui/progress-bar';
 import React from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
