+++ clayui.com/src/components/clay/Tabs.js
@@ -8,7 +8,7 @@ import ClayIcon, {ClayIconSpriteContext} from '@clayui/icon';
 import ClayTabs from '@clayui/tabs';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
