+++ clayui.com/src/components/clay/TimePicker.js
@@ -6,7 +6,7 @@
 import ClayTimePicker from '@clayui/time-picker';
 import React, {useState} from 'react';
 
-import Editor from './Editor';
+import Editor from '../Editor';
 
 const spritemap = '/images/icons/icons.svg';
